requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
cycler==0.11.0
discord.py==1.7.3
fonttools==4.33.3
greenlet==1.1.2
idna==3.3
kiwisolver==1.4.3
matplotlib==3.5.2
multidict==6.0.2
numpy==1.23.0
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.39
typing_extensions==4.2.0
wordcloud==1.8.2.2
yarl==1.7.2
